DjangoRESTFrameWork/tutorial/requirements.txt
django
raml-codec
djangorestframework
markdown       
django-filter 
django-rest-raml
pygments 